+++ plugins/hello-world/webapp/model/helloPlugin.js
@@ -3,8 +3,8 @@
  * @tagline         Hello Plugin Data Model
  * @description     Simple data model demonstrating plugin structure
  * @file            plugins/hello-world/webapp/model/helloPlugin.js
- * @version         1.7.2
- * @release         2026-07-27
+ * @version         1.7.3
+ * @release         2026-07-30
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/