+++ webapp/tests/unit/user/user-update-schema-extension.test.js
@@ -4,8 +4,8 @@
  * @description     Tests for the update() filteredData logic: regular users may now persist
  *                  schema-extension blocks where _meta.userCard.visible is true on self-update.
  * @file            webapp/tests/unit/user/user-update-schema-extension.test.js
- * @version         1.7.2
- * @release         2026-07-27
+ * @version         1.7.3
+ * @release         2026-07-30
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/